Nicer Iq Gui (#78)
* start iq_widget

* only config file changes

* started to implement iq_gui

* restores compatibility with python3

* Iq gui still needs some cosmetics but we are gettin there...

* attempt to make acbandwidths more reliable when code changes (untested)

* Iq gui is BEAUTIFUL !!!

* ComboProperties options are now defined at the FPGA level

* Nicer iq_gui

* Solve the issue raised in Pull Request #75 by Leo (float acbandwidth)

+++ pyrpl/redpitaya_modules.py
@@ -1344,7 +1344,7 @@ class IQ(FilterModule):
 
     @acbandwidth.setter
     def acbandwidth(self, val):
-        val = int(val)
+        val = float(val)
         self.inputfilter = -val
         return val
